refactor: move hugCow to its own file
