+++ src/game-logic/reducers/hugCow.test.js
@@ -1,0 +1,54 @@
+import { COW_HUG_BENEFIT } from '../../constants'
+import { generateCow } from '../../utils'
+
+import { hugCow } from './hugCow'
+
+describe('hugCow', () => {
+  describe('cow has not hit daily hug benefit limit', () => {
+    test('increases cow happiness', () => {
+      const cow = generateCow()
+      const {
+        cowInventory: [{ happiness, happinessBoostsToday }],
+      } = hugCow(
+        {
+          cowInventory: [cow],
+        },
+        cow.id
+      )
+
+      expect(happiness).toBe(COW_HUG_BENEFIT)
+      expect(happinessBoostsToday).toBe(1)
+    })
+
+    describe('cow is at max happiness', () => {
+      test('does not increase cow happiness', () => {
+        const cow = generateCow({ happiness: 1 })
+        const { cowInventory } = hugCow(
+          {
+            cowInventory: [cow],
+          },
+          cow.id
+        )
+
+        expect(cowInventory[0].happiness).toBe(1)
+      })
+    })
+  })
+
+  describe('cow has hit daily hug benefit limit', () => {
+    test('does not increase cow happiness', () => {
+      const cow = generateCow({ happiness: 0.5, happinessBoostsToday: 3 })
+      const {
+        cowInventory: [{ happiness, happinessBoostsToday }],
+      } = hugCow(
+        {
+          cowInventory: [cow],
+        },
+        cow.id
+      )
+
+      expect(happiness).toBe(0.5)
+      expect(happinessBoostsToday).toBe(3)
+    })
+  })
+})
